Increase badge cache time & add control (#1569)
Increase the default time that badges are cached
and the default time that badges are served by the CDN when stale.
Also add environment variables to control these values.

The increased default `BADGE_CACHE_MAX_AGE` (from 1 day to 10 days)
reduces server effort. Badge responses are very fast, but
it's even faster when we don't have to deal with them at all.
The current (1 day) setting means that in practice every day
we must reply a badge for every single project. We don't need to do
that; let's distribute the work further.

The increased `BADGE_CACHE_STALE_AGE` (from 10 days to 100 days)
provides us with greater protection from server problems.
If the server goes down, badges will continue to be served for a
long period of time, so many people won't even realize we're down.
We can purge them all if needed.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>