Use expires_in to show project_stats
If we just set the HTTP header 'Cache-Control', Rails quietly
adds "private" which we do *NOT* want when showing project_stats.
So, switch to using Rails' expires_in, where we can easily set
public: true.

This should significantly optimize project_stats views, because it
allows the JSON files to be cached by the CDN (as was intended).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>