Redefine CORS patterns using regexes (#1579)
The CORS patterns were defined with strings (globs). Those are
easy to read but in some cases the patterns overlapped.
The overlap meant that some .json files were being incorrectly sent
with a "Vary" over Origin, which interferes with caching.
There doesn't seem to be a way to prioritize the patterns to
prevent this as long as we stick with globs.

This commit redefines the CORS patterns with regexes.
Regexes are more complicated, but far more powerful; we can
use regexes to far more precisely declare the patterns.

As a result, the .json endpoints will no longer Vary by Origin,
which should significantly improve caching.

The final outcome is that this commit should increase average
user speed AND reduce average server load, both due to
better caching. It's always nice to get both.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>