Make app router validations pickier (#1578)
Add various validations to the router so it's pickier about
what it accepts. In particular, IDs must be positive integers.
Before we quietly converted IDs that *started* as integers
as integers, but we have attackers creatively adding garbage
after the ID to try to trigger attacks. We know of no way for
that to actually *work* as an attack on our system, but it's
better to validate input and reject anything that doesn't meet
our strict criteria. We do that in other places; now we'll make
these pickier as well.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>