Extend default FAIL2BAN pattern for % or &.
We're getting a lot of attacks of the form
/:locale/projects/:id followed by a % or & and then nonsense.
These are *legal* URLs (per the URL spec), but there's
no reason to send paths with that pattern to our site;
they simply seem to be attacks. They're naive attacks and
seem unlikely to break into the system, but since we can
detect them as attacks, let's tweak our default FAIL2BAN
pattern to counter them. This reduces the time we waste
responding to them, and might also prevent a real attack
(they might later randomly find a *working* attack, but if
we've already temporarily banned them it would be ignored).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>