Optimize project_stats CSV generation
Scout APM is reporting that our longest time reply is
project_stats when CSV format is requested.

The problem is that we first convert the *entire* project_stats
so each row is an ORM object with objects inside it, and then
generate CSVs from each. That's not necessary.

This optimizes the CSV generation
Real-world calls can take >3seconds. In benchmarks this optimization
reduced time 572ms->209ms, allocations 408339->80127.
It's the number of memory allocations that's key;
by reducing the number of memory allocations, we reduce time *and*
reduce long-term memory pressure.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>