Remove N+1 query in GitHub repo_data
This commit removes an N+1 query when we determine
the GitHub repos to display. We want to *not* show,
when asking the user what to create, the repos already
in the database. Originally we queried once for each repo
of the user; now we do it just once. It's not critical,
but it slightly speeds up project creation, and in general
we want to eliminate N+1 queries.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>