Rename "master" branch to "main" (#1586)
* Rename "master" branch to "main"

Rename the "master" branch to "main" in git.

This does *not* change the Heroku app name. That is *much*
more complicated because renaming that requires new certificates,
as well as modifications of settings in Scout APM, Heroku, and
several Heroku plugins at least, it will require new DNS entries
and a re-authentication with SendGrid, and possibly other changes
that we aren't sure we understand.

It is less risky to try to do this rename in stages, so that's
what we're doing.  In addition, the Heroku name is a lot less visible.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>