Fix remaining Ruby 3.0.0 incompatibilities
This eliminates the remaining use of ActionController::TestCase,
so we don't need its special-case login code to work
(so we can eliminate it).

This modifies the code to be easier to test. This
adds a special login mode that only works when not in production,
that enables very old session values. This eliminates the need to
rewrite the session value "remotely" for tests, as we now have an
endpoint where we can directly request an important test case value.

This eliminates the test for deny_login. The test was a problem anyway,
as it does not parallelize. We're unlikely to make changes that cause
this to fail, so it's not as important to test. We'd *like* to re-add
such a test in the future if we can figure out how to do that effectively.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>