Performance: Reduce session cookie timestamping
Up to now we have always reset `time_last_used` in
the session cookie of logged-in sessions.

This commit teaches the application to only reset
`time_last_used` if it is more than `RESET_SESSION_TIMER`
(1 hour) old.  This efficiency measure avoids constantly updating the
session cookie for many closely-related interactions (as is typical).

This should significantly reduce the number
of session cookie settings we send back to the requestor.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>