Add "maintained" criterion
This adds a new "maintained" criteria, per mailing list discussion.

We normally only add criteria after a long drawn-out time, because
we don't want people to lose their earned badges without giving them
lots time to fix things.

However, this case is diffeerent.
We have *always* assumed that a project actively pursuing a badge
is maintained. In this case we are simply adding a criterion we always
assumed but failed to properly document. Since we always assumed it, we can
set maintained=Met by default. With this default, existing projects will
*not* be at risk of losing a badge, so the reason we normally are
slow about adding criteria does not apply in this special case.
There seems to be unanimous agreement on the CII Best Practices mailing
list that it'd be appropriate to add this criterion.

In particular, it seems inappropriate that a project that is *known*
to be unmaintained should retain a best practices badge.
As Daniel Stenberg noted on 2021-03-24,
'Marking a project as "unmaintained" should disqualify it
from following best practices. Unmaintained is a *worst* practice'.

This commit does *NOT* automate anything. That will be handled later and
separately. There are two kinds of "unmaintained": projects that
*expressly* state they are unmaintained, and projects where we can
infer that they are probably unmaintained because they are large and inactive.
This automation will be handled later; currently we just want
to make sure the criterion is *present*. Adding this criterion
immediately allows us to mark "maintained=Unmet" on a few projects where
we know this criterion is not met.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

Fix tests given new criterion 'maintained'

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>