Make percentage <100% if level incomplete
This commit caps the "percentage" value to
99 at most if the level is not 100% complete.

This is a defensive programming measure.
We round percentages, so this will only a problem once
(N-1)/N >= 0.995, that is, number of items in a level N>=200,
and we don't want that many questions in a level.
Still, I feel better being *certain* it cannot cause problems.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>