Add script to programmatically modify project
This is a simple Python script to programmatically
modify a project on the Best Practices badge site.
This is in Python3 because the original requestor wanted
it in Python3.  It should be easy to port to other languages,
*or* just execute it.

I had trouble making it work when using PATCH on a JSON
endpoint, so this uses HTML forms.
There were several knotty little cases that needed to be worked out.

It create a module docstring & use it also as the "--help"
epilog. That provides help information in two places
*without* duplicating the text. My thanks to TonyLHansen
for the recommendation.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>