DO NOT MERGE: Add demo script to modify project
This is starting work to create a simple Python script
to programmatically modify a project on the Best Practices badge site.

I had trouble making this work earlier, so I'm starting over here
and making it more visible so others can see.

This is in Python3 because the original requestor wanted it in Python3.
It should be easy to port to other languages, *or* just
execute it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>