Modify api.md to point to Python files
Modify api.md. We've renamed one Python file, so we need
to change the link to it.

In addition, we've added a new Python script to show how to
modify project data, so let's add a link from api.md
to this new Python script showing how to do that.
This should make it easier for people to programmatically
update data as appropriate.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>