Add script to programmatically modify project (#1597)
* Add script to programmatically modify project

This is a simple Python script to programmatically
modify a project on the Best Practices badge site.
This is in Python3 because the original requestor wanted
it in Python3.  It should be easy to port to other languages,
*or* just execute it.

I had trouble making it work when using PATCH on a JSON
endpoint, so this uses HTML forms.
There were several knotty little cases that needed to be worked out.

It create a module docstring & use it also as the "--help"
epilog. That provides help information in two places
*without* duplicating the text. My thanks to TonyLHansen
for the recommendation.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Use "is None" not "== None"

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Rename Python scripts to remove '-'

It's challenging for Python to import filenames with '-',
so rename these Python files to use '_' instead.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Modify api.md to point to Python files

Modify api.md. We've renamed one Python file, so we need
to change the link to it.

In addition, we've added a new Python script to show how to
modify project data, so let's add a link from api.md
to this new Python script showing how to do that.
This should make it easier for people to programmatically
update data as appropriate.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Improve help documentation

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>