Update Fastly purge_all
Update `rake fastly:purge_all` to work properly.

We previously had to remove the FastlyRails gem, but
never changed how to do a `purge_all`, so that rake command
no longer works.

This commit re-implements purge_all so that it works again
(and renames it from `purge` to `purge_all` to clarify that
is what it's doing). We must `purge_all` when, for example,
there's a change in how the badges are calculated (so that
recipients will see the new results).

Sadly, due to a limitation of Fastly's API,
we *cannot* do a soft purge, aka "Fastly-Soft-Purge: 1".
A soft purge would let the CDN distribute the cached value
if it couldn't contact the real server. Unfortunately,
Fastly doesn't support soft purges on full purges.

Beware: Do *not* run `purge_all` lightly on real systems.
It will temporarily increase server load, because it flushes
*EVERYTHING* that was cached on Fastly. As a result,
the server will have to provide the cached data.
But when you need it, you need it.

There's some rigging of the test (as we've done with other
Fastly method calls), because we don't want tests to
*actually* contact the Fastly service :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>