Fix purge_all (key != API key)
Fix Fastly purge_all. There are two different keys normally
used in purges, the key of the particular value to be purged
AND the API key. In `purge_all`, however, there's no need for
the first key (we are purging EVERYTHING), so the "normal"
key should not be a parameter of the `purge_all` call.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>