test/integration/recalc_test.rb: Reload @projects
Test 'Make sure recalc percentages only updates levels affected'
is flapping (sometimes fails and sometimes doesn't).
It *only* flaps on CircleCI, not in local tests, which makes it
very hard to debug.

This commit forcibly reloads @project at the beginning
of the two tests in test/integration/recalc_test.rb
that use @project. This isn't a bad idea anyway, and we
hope this fixes the problem.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>