Add rake task "search_user" to simplify GDPR requests
When the LF gets GDPR requests to erase user data, we
must search the CII Best Practices database.

This commit makes searching simpler and faster, by creating a new
`rake search_user NAME EMAIL` command that can search for
*both* name and email in a single command. Most of the time we
get a single name and email address in each request, and it takes
a while to spin up the rake task, so this way we get all the answers
for a single user at once.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>