Update gem puma (4.3.6->5.2.2) (#1620)
Update gem puma, our web server. This is a trivial change
to the Gemfile (and to the generated file Gemfile.lock).

I read through the puma 5 upgrade section before doing this.
There are several changes to its API and defaults, but I didn't
see any changes that affect this application. For example,
the default number of threads changed, but we force a specific
value for that, so that change shouldn't matter to us.

All tests pass after upgrading.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>