Move `access_continuity` tweak into en.yml
Cary Phillips (@cary-ilm) proposed a nice improvement to the
text of criterion `access_continuity`. It changes
"incapacitated or killed" to the better (and intended)
"dies, is incapacitated, or is otherwise unable or unwilling to
continue support".

However, the previous implementation
modfied the *generated* file `doc/other.md` instead of the
source file `config/locales/en.yml`.

This commit moves the criterion text change into the generated
file. That way, this improvement will be used and translated.

This commit, along with the ones it builds on, resolves PR #1625.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>