Note OpenSSF!
Note that this project is now part of the OpenSSF.
The CII officially has ended, and it's been succeeded by the
Open Source Security Foundation (OpenSSF).

I meant to do this earlier, but in the process of adding information
on the CII website I noticed that this hadn't been done.
So, let's fix it!

For some people this may not *matter*, but it's generally good
to note "where this project fits in" where it's part of
something larger.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>