Update transitive dependencies (bundle update)
Run "bundle update" to update transitive dependencies.
This updates many gems, including "addressable".

This technically eliminates a vulnerability in the gem "addressable",
namely CVE-2021-32740, "Regular Expression Denial of Service in
Addressable templates", More information is here:
https://github.com/advisories/GHSA-jxhc-q857-3j6g

CVE-2021-32740 is highly unlikely to be exploitable in our software,
as we don't allow untrusted users to set or edit templates.
Still, it's easier to update than to ensure that we're not
vulnerable, and safer as well. So by policy we update
unless there's a reason not to update.

Note that this change incorporates pull request
<https://github.com/coreinfrastructure/best-practices-badge/pull/1637>.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>