know_common_errors: edX course is free
Clarify in `know_common_errors` (in addition to `know_secure_design`)
that the edX course is free, while the certificate is not.
This implements the intent of #1643.

Previous commit b3063cf87df66ed51e98f6de25e38797740b06bf
clarified that taking the edX course is free, while earning
the certificate is not.  edX is emphasizing the pay option of getting
an edX course so much that it's not obvious when the course itself
is free (as it is in the case of the fundamentals courses).

That previous commit changed the details for `know_secure_design`,
but it didn't fix the same text in `know_common_errors`.
This commit fixes the equivalent text in `know_common_errors`.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>