dockerfiles/README.md: Clarify update is required
Tweak the dockerfiles/README.md file to make it clear that you MUST
update .circleci/config.yml when you update a docker image.  If you
don't update the CircleCI configuration, then the system will continue
to use the old image. That's as intended; the whole point of pinning is
to only use the image specifically identified.

This is a side-effect of pinning image dependencies.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>