Add .github/.dependabot.yml
It's possible to enable dependabot via just GitHub project settings,
and we did that for a while. However, when a project does that,
it's harder for others (such as OpenSSF Scorecard) to realize or verify
that the project has dependabot enabled. Creating this dependabot.yml
file makes easier to *verify* that automatic dependency update proposals
are happening. This also gives finer-grained control.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>