Fix source code formatting per rubocop 1.22.1
This commit will help us eventually update rubocop.

Currently we can't update rubocop easily because it reports many issues.
This commit is the result of all automated fixes when you update to
rubocop 1.22.1, set LineLength to 79, and use `rubocop -a`.
In a few cases I had to hand-fix the formatting.

They're all trivial changes, mainly changes in indentation, reversals
of conditionals (to avoid the use of negation in the condition), and
elimination of unnecessary key quotation.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>