Update to rails 6.1.4.4, warn re: sprockets-rails
Update to rails version 6.1.4.4 and warn about updating
the gem sprocket-rails.

The gem sprockets-rails, when updated, works fine locally and
passes all tests, but mysteriously the icons look corrupted.
We'll eventually need to figure that out, but our priority
is updating rails (just in case the known vulnerability is
exploitable, even though we don't think it is).
We also need to update the name (finally), and that also
takes priority over updating sprockets-rails.
So we'll *clearly* document the issue with a warning in the Gemfile,
so we won't re-trigger the problem accidentally and document
hints to fix it when the time comes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>