Simplify SVG display of Hyperledger logo
It turns out that you can just use "img src" on
SVGs just as easily as PNGs. I thought Rails didn't
support it, but it does; I think I made a mistake
when I tested it.

So this turns out to be really easy, I just took the
hard way to get to the easy way :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>