Bump rubocop from 0.91.1 to 1.0.0
This required changing URI.open(...) to URI.parse(...).open(...).
The URI.open() operator is (like Perl) dangerously magical; a leading
'|' will be specially interpreted. This isn't a security vulnerability
in our case because our validators prevent the dangerous case, but it's
much safer to change the code so that even if the validators fail or
get circumvented an attacker *still* can't cause the dangerous behavior.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>