Add "blocked" field to users table
This adds "blocked" and "blocked_rationale" fields to the users table.

Unfortunately some users choose to create fraudulent entries.
They're rare, but when it happens we want to block them
from being able to change anything, and record why.

This commit *only* modifies the schema so we can record this
information - it does not *enforce* them. I'm creating them as
separate commits so that we can record this information immediately.
We can then later add the code to use this information to enforce it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>