Bump Rails to 6.1.4.7
Bump version of Rails to 6.1.4.7.
In particular this updates for vulnerability
CVE-2022-21831, GHSA ID GHSA-w749-p3v6-hccq,
a high-severity vulnerability of weakness class CWE-94.
It's not clear the site is directly exploitable with it,
but our polity is to simply update instead of wasting time
doing a deep analysis to figure out if it's exploitable in our case.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>