Update transitive dependencies (`bundle update`)
Update the transitive dependencies by running `bundle update`.

This bumps `rack` from 2.2.3 to 2.2.3.1, repairing
[CVE-2022-30123](https://github.com/advisories/GHSA-wq4h-7r42-5hrr).
This is a shell escape sequence injection vulnerability in the Lint
and CommonLogger components of Rack.
The problem is that carefully crafted requests can cause shell escape
sequences to be written to the terminal via Rack's Lint middleware and
CommonLogger middleware. These escape sequences can be leveraged to
possibly execute commands in the victim's terminal.

We probably aren't vulnerable to this. Known attacks only apply when
`Rack::Lint` or `Rack::CommonLogger` are used, and a `grep -R`
finds neither. However, our general policy is to simply update
if there's a security vulnerability in our components and the update
passes our tests. It's possible that it's being used
(e.g., via one of the libraries we import or via dynamically created
dependencies). The risk is very low in updating for a security vulnerability
as long as our tests pass, and we believe the risk is *higher* if
we don't update. We're prepared to update quickly, so we do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>