Begin transition to SACM notation
This begins (but does not complete) a transition from
CAE notation to SACM notation. The idea is that we'll complete
the transition over time.

* Add early version of assurance-case-verification.odg

This adds an early version of assurance-case-verification.odg.

Scott Ankrum is providing useful feedback to improve the
CII Best Practices assurances case. His goal is to create
a SACM version of the assurance case, using a SACM tool.
In the process, he's provided a lot of feedback on how to
improve our assurance case.

I haven't committed to switching from CAE notation to SACM notation.
Also, the tool they're using does not appear to be open source software
(a requirement on all tools needed to modify the BadgeApp)
and seems to be alpha-quality early stage software.
However, I definitely want to take advantage of his insights.

So I intend to work with him to modify our current assurance case
information, so that our assurance case in CAE notation can be
trivially converted to SACM and back.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* assurance-case-verification.org: Updates

Add evidence. Change icon fills to more closely match:
https://www.adelard.com/asce/choosing-asce/cae.html
The icons shown are all white with only colored edges,
but the colors are hard to see; changing the icons
to be fully colored makes the color differences easier to see.
We continue to use their icons.
So this means we have

* oval (with blue fill) = claim/subclaim
* rounded rectangle (with green fill) = argument
* box (with purple fill) = evidence

Fills are intentionally light so they don't interfere with
reading the black text.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Switch top-level to mostly claims

Scott Ankrum (MITRE) has been redoing the assurance case
and strongly recommends rewording things as a sequence of
subclaims, and is recreating the assurance case in SACM.
It *is* easier to understand as subclaims.

This commit revises the top level as claims and sub-claims.
It follows many of his wordings, but with some changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak assurance case text & id some negative tests

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Note GDPR document and not showing email addresses

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak security.md

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update some diagrams per Scott Ankrum feedback

In particular, switch to using claims / sub-claims in many cases.
Since claims are true/false statements, they can be easier to
understand.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add missing evidence

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add SACM version of assurance-case-verification

I tried using the SACM graphical notation instead of CAE.
This experiment showed that SACM has some distinct advantages!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* In SACM use more bigdots

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Small tweaks to top-level assurance case diagram

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update assurance case verification (SACM and not)

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add assurance-case-verification-cae.png

Add a PNG file that unambiguously shows a CAE version,
to make it easier for others to discuss this.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add draft text justifying use of SACM

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update .odg files

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update assurance case text - explain SACM

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak for asCited

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak security.md

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Proposed revised toplevel in SACM format

This proposed toplevel reorganizes some things, so that this
toplevel is easy to reuse in other systems.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Simplify top level

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Change evidence fill color to light magenta 4

Fill colors make it even easier to see the icon differences
(if you can see color). We've been using the CAE edge colors
as the fill colors. For evidence, switch to light magenta 4.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Show SACM notation in SVG

Show the SACM notation we use.

GitHub can now include SVG, so use SVG as that produces
a better display.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* design.md: Cite another document about our approach

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update top level SACM version

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add reference to new SACM top level

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update SACM notation sample

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Extend SACM notation list

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak SACM notation summary further

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update toplevel, clarify we don't use waterfall

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Note "physics envy"

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak CAE/SACM discussion

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>