Update to Rails 6.1.6.1
Update all Rails components to 6.1.6.1.
This addresses CVE-2022-32224. I don't think this is exploitable
in our circumstance, because we don't deserialize YAML from a database.
That said, updating is the less risky course.

This also updates octokit, this addresses CVE-2022-31072.
Again, this shouldn't be vulnerable - that is a bad file permission
that shouldn't matter because we don't give external access to the
system's files.

This adds faraday-retry as a dependency. These other updates update
the gem faraday. Its retry abilities are now in a separate gem,
faraday-retry. We *do* want to enable retries, because that provides
better reliability for users.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>