When creating new local account show static reply
Modify `app/controllers/users_controller.rb` so that
when a local user account is created it always presents
the same reply. If we don't, attackers could determine if
a given email account *exists* as a local account.
They still couldn't figure out which account, or practically
enumerate the addresses present, but it'd be best
to not even let attackers do *that*.

As before:
- If someone tries to create a local account that currently doesn't
  exist, the local account will be created but not be activated
  (and we will send the email to explain how to activate it).
- If someone tries to create an account that exists but the account
  is not activated, we will resend the activation email to the user account.

Other than that, we quietly do nothing. That way, valid users will
get the emails they're entitled to (at their email address), but
otherwise we don't bother anyone. In particular, once the account
is activated, there is no reason to resend activation emails - so we don't.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>