Make password resets *always* reply same message
Previously we replied with error reports if there was no account or if
there was a GitHub account (not a local account) with the email address.
However, that allowed attackers to determine if a given email address
was present or not in an account. That's not a large leak of
information; the attacker has to already guess the specific email address,
and they merely get "is present/absent" instead of the specific account.
Many systems *do* provide such error messages, so many users would
be unsurprised by this information leak.
Still, we want to be excellent at providing user privacy, so we're
going to go beyond what some might see as the minimum, and instead
do what we can to maximize our users' privacy.

We previously made a similar change to prevent leaking the presence
or absence of an email address when creating local accounts.

This change modifies the tests to match. While we're at it, we'll
remove a call to the deprecated assigns() helper in the test.

Note: This eliminates some keys for translation, as well as mildly
tweaking the text for a key.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>