Update to OpenSSF Code of Conduct (CoC)
We are now an OpenSSF project, so we need to apply the OpenSSF
Code of Conduct (CoC). This project predates the OpenSSF, and when we
added a CoC we used a different CoC from the Linux Foundation.
Now that we *are* part of the OpenSSF, we should have update this. However,
I just realized that I never updated the CoC to match. This fixes that.

We generally haven't had Code of Conduct problems, and the core
*fundamentals* are unchanged: treat everyone else well.
Still, I think it's best to prepare to counter potential problems
ahead-of-time.

The markdown text was created by downloading the text from
<https://openssf.org/community/code-of-conduct/> and then using the
online tool <https://codebeautify.org/html-to-markdown>.
This file *does* link to the OpenSSF CoC page, but I thought it'd be
wise to include a markdown copy of the text so that there was *no*
question about what it said.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>