Remove `main` tier for deployment
Heroku has eliminated its free services, which we originally used
for both our main/master tier and staging tier.

This commit removes the "main/master" tier as a cost-cutting move.
We'll still have the "staging" and "production" tiers; use
`rake deploy_staging` and `rake deploy_production`  to deploy
to those respective tiers. This eliminates the attempt to
deploy to the "main/master" tier (which is about to be deleted),
as well as updating the documents to reflect this change.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>