Change how feed view is linked
I am trying to determine why users are being redirected when they
should *not* be redirected.

When viewing the main page, the URLs are fine until
"auto_discovery_link_tag" is called, and from then on the domain
name is wrong. My hypothesis: this routine is incorrectly setting
a global variable for the hostname, causing later calls to be wrong.

So let's instead force the URL to a local reference, which is fine
and eliminates the call entirely. I don't know if this will solve the
problem, but this change is harmless, and its results can tell us
if we're on the right track.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>