Fix email (#2054)
* Modify email sending configuration

We need to move off SendGrid. First step: modify how we send email
so it's much easier to configure.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix documentation so it doesn't say we use SendGrid

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add link to Amazon SES docs

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Send email using TLS directly (without STARTTLS)

Switch to sending email using TLS directly, without using STARTTLS at all.

STARTTLS lets us request an upgrade to TLS, and we *could* choose to not
send an email if the upgrade fails.

However, it's simpler to simply use TLS right from the beginning. Amazon
supports this, and many others probably do as well. This is simpler, and thus
less prone to failures if an "upgrade to TLS" fails. There's never an opportunity
for any party to be "confused" about whether or not TLS is currently enabled.

Of course, this *only* applies to the hop from our application to the MTA
who will then forward the email on. MTAs communicate with each other in a variety
of ways. But if we can securely get the email to an MTA we trust, the MTA can
then forward the email using the best mechanism available to it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>