Create LLC and TSC
Clean up various legal/charter constructs.

The Linux Foundation plans to move this project into its own LLC
(this provides various legal protections), and we must state that.

A side-effect is that we need to create "Technical Steering Committee"
(TSC). I think that's a good thing. I intend to start it with me
(as I was the original project lead & committer), and work to add
others to it, so it's not just me working by myself.

The proposed approach says the OpenSSF TAC can change the TSC
by 2/3rds vote. The intent is to ensure that I (or anyone else) can't
be a dictator, and that it's possible to get things going again
if something goes wrong. Originally I was going to have the
OpenSSF Best Practices WG do that vote, but I think it'd be
better to have the TAC do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>